Refactor url generations
Move the responsibility to generate url values from
org.javaunit.autoparams package to org.javaunit.autoparams.generator
package.
