--- autoparams/src/main/java/org/javaunit/autoparams/SimpleValueObjectGenerator.java
@@ -1,13 +1,0 @@
-package org.javaunit.autoparams;
-
-@Deprecated
-@SuppressWarnings("DeprecatedIsStillUsed")
-final class SimpleValueObjectGenerator extends CompositeObjectGenerator {
-
-    public SimpleValueObjectGenerator() {
-        super(
-            new UrlGenerator() // TODO: Move to the new simple value generator.
-        );
-    }
-
-}
