+++ autoparams/src/main/java/org/javaunit/autoparams/generator/SimpleValueObjectGenerator.java
@@ -21,7 +21,8 @@ final class SimpleValueObjectGenerator extends CompositeObjectGenerator {
             new TypeMatchingGenerator(Factories::createLocalTime, LocalTime.class),
             new TypeMatchingGenerator(Factories::createLocalDateTime, LocalDateTime.class),
             new TypeMatchingGenerator(Factories::createClass, Class.class),
-            new EnumGenerator()
+            new EnumGenerator(),
+            new UrlGenerator()
         );
     }
